start CPU rewrite
